refactor: extract useOutsideClick as separate module
#72
